requirements.txt
Babel==1.3
Chameleon==2.14
Mako==0.9.1
MarkupSafe==0.18
PasteDeploy==1.5.2
Pillow==2.3.0
SQLAlchemy==0.9.3
WebHelpers2==2.0b5
WebOb==1.3.1
WebTest==2.0.14
beautifulsoup4==4.3.2
colander==1.0b1
colorama==0.2.7
deform==2.0a2
peppercorn==0.4
py==1.4.20
pyramid==1.5b1
pyramid-deform==0.2
pyramid-layout==0.8
pyramid-mailer==0.13
pyramid-mako==0.3.1
pyramid-tm==0.7
pytest==2.5.2
pytz==2013.9
repoze.lru==0.6
repoze.sendmail==4.2
six==1.5.2
testfixtures==3.0.1
transaction==1.4.1
translationstring==1.1
venusian==1.0a8
waitress==0.8.8
zope.deprecation==4.1.0
zope.interface==4.1.0
zope.sqlalchemy==0.7.4
